+++ src/main/java/org/bot/nullbot/command/reply/HelpCommand.java
@@ -7,9 +7,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.bot.nullbot.annotation.CommandMapping;
 import org.bot.nullbot.command.Command;
+import org.bot.nullbot.component.resource.ResourceLoader;
 import org.bot.nullbot.config.FileStorageConfig;
 import org.bot.nullbot.entity.CommandEvent;
-import org.bot.nullbot.util.ResourceLoader;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
@@ -21,12 +21,13 @@ import java.io.IOException;
 public class HelpCommand implements Command
 {
     private final FileStorageConfig fileStorageConfig;
+    private final ResourceLoader resourceLoader;
 
     @Override
     public void execute(Bot bot, CommandEvent<?> event) {
         if (event.getEvent() instanceof GroupMessageEvent groupMessageEvent) {
             try {
-                String helpPath = ResourceLoader.getCached("static/help/help.jpg", fileStorageConfig.getTempPath()).toAbsolutePath().toString();
+                String helpPath = resourceLoader.getCached("static/help/help.jpg", fileStorageConfig.getTempPath()).toAbsolutePath().toString();
                 String response = MsgUtils.builder().img(helpPath).build();
                 bot.sendGroupMsg(groupMessageEvent.getGroupId(), response, false);
                 log.info("\t\t\t\t├─[Help] 已获取帮助");